fix: change existing_customer_id to unsignedBigInteger for foreign key compatibility
- customers.id is bigint unsigned
- existing_customer_id must match to create foreign key constraint

+++ database/migrations/2025_07_26_001100_create_import_temp_customers_table.php
@@ -50,7 +50,7 @@ return new class extends Migration
             // References
             $table->unsignedBigInteger('import_job_id');
             $table->foreign('import_job_id')->references('id')->on('import_jobs')->onDelete('cascade');
-            $table->unsignedInteger('existing_customer_id')->nullable(); // Link to existing customer if duplicate
+            $table->unsignedBigInteger('existing_customer_id')->nullable(); // Link to existing customer if duplicate
             $table->foreign('existing_customer_id')->references('id')->on('customers')->onDelete('set null');
             
             $table->timestamps();
